Adding some basic administrative functionality:
Cleaned up some bootstrap.php code, in general, and added a service to deliver the pagination handler.
Deleted the AdminBaseController.
Added route and methods to display recipes by session user.
Added the admin template userRecipeList.
Removed css property for img max-width 100%.
Fixed login bug in the script.js file, it was appending the login dashboard to the current url, not replacing the path.